+++ example/linux/flutter/generated_plugins.cmake
@@ -3,8 +3,8 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
-  url_launcher_linux
   maps_launcher
+  url_launcher_linux
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)