+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -1,0 +1,17 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#include "generated_plugin_registrant.h"
+
+#include <maps_launcher/maps_launcher_plugin.h>
+#include <url_launcher_windows/url_launcher_plugin.h>
+
+void RegisterPlugins(flutter::PluginRegistry* registry) {
+  MapsLauncherPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("MapsLauncherPlugin"));
+  UrlLauncherPluginRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("UrlLauncherPlugin"));
+}